fix: Resolve clippy warnings for all-targets all-features
- Remove assert!(true) statements that get optimized out (use implicit pass)
- Remove unused import TracingGuard in tracing_init_test.rs
- Remove unused import Throughput in zero_copy_benchmark.rs
- Fix unnecessary cast in tracing_performance_test.rs
- Fix unused Result in benchmark by using let _ = pattern

All clippy checks now pass with: cargo clippy --all-targets --all-features -- -D warnings

+++ tests/tracing_performance_test.rs
@@ -39,7 +39,7 @@ fn test_extract_trace_context_performance() {
     }
     let elapsed = start.elapsed();
 
-    let avg_nanos = elapsed.as_nanos() / iterations as u128;
+    let avg_nanos = elapsed.as_nanos() / iterations;
     println!("Average extract time: {}ns", avg_nanos);
 
     // Should be less than 3μs (3000ns) - realistic target for HashMap lookup + parsing
